feat(security) block new registrations
This commit introduces a new trigger in the token database to block new accounts registrations. It's a workaround needed to avoid that anyone with network access to the service can use it. The value can be tuned changing the `MAX_USERS` variable and restarting the `tokenserver_db_init` container.

Minor secondary improvements:
- added permissions to the GitHub workflow
- added gitignore
- switched MariaDB images to LTS
- minor fixes to the readme
- separate user and password for the syncstorage and tokenserver databases
- add a script to create random passwords and initialize the .env and example files