chore(docker) move syncstorage init to different image
THIS IS A BREAKING CHANGE: docker-compose.yml and the .env files need to be updated
Previously, the syncstorage-rs base image was altered to include the mariadb client and the dbinit.sh script. Now, a new image from mariadb:lts is built with the initialization script inside. This reduce the total space used by the docker layers, reduce the complexity of the code to maintain and adds more transparency about changes to the base images.